+++ Source/Lib/Fluxor.Blazor.Web.ReduxDevTools/Internal/ReduxDevToolsInterop.cs
@@ -1,20 +1,33 @@
-﻿using Fluxor.Blazor.Web.ReduxDevTools.CallbackObjects;
+﻿using Fluxor.Blazor.Web.ReduxDevTools.Internal.CallbackObjects;
 using Microsoft.JSInterop;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Fluxor.Blazor.Web.ReduxDevTools
+namespace Fluxor.Blazor.Web.ReduxDevTools.Internal
 {
+	public interface IReduxDevToolsInterop
+	{
+		bool DevToolsBrowserPluginDetected { get; }
+		Func<JumpToStateCallback, Task> OnJumpToState { get; set; }
+		Func<Task> OnCommit { get; set; }
+		ValueTask InitializeAsync(IDictionary<string, object> state);
+		Task<object> DispatchAsync(
+			object action,
+			IDictionary<string, object> state,
+			string stackTrace);
+		Task DevToolsCallback(string messageAsJson);
+	}
+
 	/// <summary>
 	/// Interop for dev tools
 	/// </summary>
-	internal sealed class ReduxDevToolsInterop : IDisposable
+	internal sealed class ReduxDevToolsInterop : IDisposable, IReduxDevToolsInterop
 	{
 		public const string DevToolsCallbackId = "DevToolsCallback";
 		public bool DevToolsBrowserPluginDetected { get; private set; }
-		public Func<JumpToStateCallback, Task> OnJumpToState;
-		public Func<Task> OnCommit;
+		public Func<JumpToStateCallback, Task> OnJumpToState { get; set; }
+		public Func<Task> OnCommit { get; set; }
 
 		private const string FluxorDevToolsId = "__FluxorDevTools__";
 		private const string FromJsDevToolsDetectedActionTypeName = "detected";
@@ -39,7 +52,7 @@ namespace Fluxor.Blazor.Web.ReduxDevTools
 			DotNetRef = DotNetObjectReference.Create(this);
 		}
 
-		internal async ValueTask InitializeAsync(IDictionary<string, object> state)
+		public async ValueTask InitializeAsync(IDictionary<string, object> state)
 		{
 			IsInitializing = true;
 			try
@@ -52,7 +65,7 @@ namespace Fluxor.Blazor.Web.ReduxDevTools
 			}
 		}
 
-		internal async Task<object> DispatchAsync(
+		public async Task<object> DispatchAsync(
 			object action,
 			IDictionary<string, object> state,
 			string stackTrace)
@@ -113,9 +126,9 @@ namespace Fluxor.Blazor.Web.ReduxDevTools
 		}
 
 		private static bool IsDotNetReferenceObject(object x) =>
-			(x is not null)
-			&& (x.GetType().IsGenericType)
-			&& (x.GetType().GetGenericTypeDefinition() == typeof(DotNetObjectReference<>));
+			x is not null
+			&& x.GetType().IsGenericType
+			&& x.GetType().GetGenericTypeDefinition() == typeof(DotNetObjectReference<>);
 
 		private ValueTask<TResult> InvokeFluxorDevToolsMethodAsync<TResult>(string identifier, params object[] args)
 		{
@@ -167,7 +180,7 @@ window.{FluxorDevToolsId} = new (function() {{
 				// Notify Fluxor of the presence of the browser plugin
 				const detectedMessage = {{
 					payload: {{
-						type: '{ReduxDevToolsInterop.FromJsDevToolsDetectedActionTypeName}'
+						type: '{FromJsDevToolsDetectedActionTypeName}'
 					}}
 				}};
 				const detectedMessageAsJson = JSON.stringify(detectedMessage);