+++ Source/Tests/Fluxor.UnitTests/IsExternalInit.cs
@@ -1,0 +1,9 @@
+﻿#if NETCOREAPP3_1
+using System.ComponentModel;
+
+namespace System.Runtime.CompilerServices
+{
+	[EditorBrowsable(EditorBrowsableState.Never)]
+	internal class IsExternalInit { }
+}
+#endif